third_party/blink/web_tests/external/wpt/app-history/navigate/navigate-in-initial-about-blank-document-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<body>
<script>
promise_test(async t => {
  let i = document.createElement("iframe");
  document.body.appendChild(i);
  i.contentWindow.appHistory.onnavigate = t.unreached_func('onnavigate should not be called');
  i.contentWindow.appHistory.onnavigatesuccess = t.unreached_func('onnavigatesuccess should not be called');
  i.contentWindow.appHistory.onnavigateerror = t.unreached_func('onnavigateerror should not be called');
  let promise_resolved = false;
  i.contentWindow.appHistory.navigate(new URL("/common/blank.html?1", location.href).href)
      .then(() => promise_resolved = true);
  await new Promise(resolve => i.onload = resolve);
  assert_equals(i.contentWindow.location.search, "?1");
  assert_false(promise_resolved);
}, "navigate in initial about:blank document");
</script>
</body>
